test: MajorQuestionUpdate,Make Controller 테스트 수정
