+++ persistence/src/main/java/com/comssa/persistence/question/repository/querydsl/LicenseMultipleChoiceQuestionDslRepository.java
@@ -4,7 +4,6 @@ package com.comssa.persistence.question.repository.querydsl;
 import com.comssa.persistence.question.domain.common.QuestionCategory;
 import com.comssa.persistence.question.domain.license.LicenseMultipleChoiceQuestion;
 import com.comssa.persistence.question.domain.license.QLicenseMultipleChoiceQuestion;
-import com.comssa.persistence.question.repository.querydsl.booleanbuilder.LevelsAndCategoryBooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.stereotype.Repository;
