+++ persistence/src/main/java/com/comssa/persistence/question/repository/querydsl/MajorDescriptiveDslRepository.java
@@ -4,7 +4,6 @@ import com.comssa.persistence.question.domain.common.QuestionCategory;
 import com.comssa.persistence.question.domain.common.QuestionLevel;
 import com.comssa.persistence.question.domain.major.MajorDescriptiveQuestion;
 import com.comssa.persistence.question.domain.major.QMajorDescriptiveQuestion;
-import com.comssa.persistence.question.repository.querydsl.booleanbuilder.LevelsAndCategoryBooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.springframework.stereotype.Service;
 
@@ -24,7 +23,6 @@ public class MajorDescriptiveDslRepository
 		List<QuestionCategory> questionCategories,
 		List<QuestionLevel> questionLevels,
 		boolean approved) {
-
 		return getQuery(question, whereCategoriesAndLevels(
 			question._super,
 			questionCategories,
